Backend/requirements.txt
psycopg==3.2.3
python-dotenv==1.0.1
Flask==3.0.3